# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OweHost API
  description: |
    OweHost Backend API - A comprehensive hosting panel backend system.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT Bearer Token**: Use the `/api/v1/auth/login` endpoint to obtain tokens
    2. **API Key**: Use the `ApiKey` prefix in the Authorization header
    
    ## Rate Limiting
    
    API requests are rate limited. Check response headers for limit information.
    
  version: 1.0.0
  contact:
    name: OweHost Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/domains:
    get:
      tags:
        - Domains
      summary: List domains
      operationId: listDomains
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of domains
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Domain'
    post:
      tags:
        - Domains
      summary: Create domain
      operationId: createDomain
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainCreateRequest'
      responses:
        '201':
          description: Domain created
  /api/v1/domains/{id}:
    get:
      tags:
        - Domains
      summary: Get domain by ID
      operationId: getDomain
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Domain details
    delete:
      tags:
        - Domains
      summary: Delete domain
      operationId: deleteDomain
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Domain deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    Domain:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [primary, addon, parked, alias]
        status:
          type: string
          enum: [active, pending, suspended]
        document_root:
          type: string
        validated:
          type: boolean
        created_at:
          type: string
          format: date-time
    DomainCreateRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum: [primary, addon, parked, alias]
        document_root:
          type: string
        target_domain:
          type: string
...
security:
  - bearerAuth: []